Option to not change wd in exec-as-user sample
Executing /usr/bin/exec-as-user user . command [args] -- the '.' there
prevents chdir to user's home directory. Alternative would have been to
use '~' to do the chdir and not chdir by default -- but ~ is expanded
by shell when quoted, therefore more error prone.

Also noted in GNUmakefile, that all in the current directory is sent
(unnecessarily) to docker daemon during docker build and this is
unavoidable (without copying files around, for the time being).